intermediate checkin
