+++ backend/data_tools/pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
 pythonpath = .
 testpaths = tests
-addopts = --cov
